requirements.txt
Flask==2.0.1
tensorflow==2.6.0
pandas==1.2.5
numpy==1.19.5
Pillow==7.2.0
plotly==4.14.3
gunicorn==20.1.0